fix(uplc): Data::Constr tag is arbitrary-precision BigInt, not u64 (#859, closes #828 residual)
Haskell's `Data = Constr Integer [Data] | ...` uses an unbounded signed
Integer tag. dugite's `Data::Constr(u64, Vec<Data>)` could not represent
a transient negative or >u64::MAX tag that a script can legitimately
build via `constrData` at protocol versions before the van Rossem
(PV11) Word64-unlifting gate, so dugite errored where Haskell succeeds.

- Widen `Data::Constr`'s tag to `num_bigint::BigInt`, matching `Data::I`.
  Update all trait impls (Clone/Drop/Hash/PartialEq/into_constr) and the
  CBOR encoder/decoder accordingly; fix a latent BigInt leak in the
  iterative `Drop` impl (the Constr tag field was never explicitly
  dropped, only the args Vec was).
- `encode_constr` now takes `&BigInt`: the compact tag-121..127 /
  tag-1280..1400 forms still apply only to `[0, 127]`; anything else
  (including an out-of-range tag) falls through to the general tag-102
  form using the same arbitrary-precision `encode_integer` used for
  `Data::I` (byte-identical for in-range tags, "soft"
  bignum/negint degrade for out-of-range ones — matching Haskell's
  unbounded `encodeInteger`). Decode still always reads a Word64 (a wide
  tag can never be on-chain-decodable), so a wide-tag CBOR blob is a
  one-way, non-round-tripping encode only.
- `constrData` denotation: gate the Word64 range check to
  semantics-variant D/E (PV>=11) only, matching Haskell exactly; A/B/C
  (pre-PV11) now accept ANY BigInt tag with no representational gap and
  no `Internal` fallback. `unConstrData` passes the tag through as-is.
- Textual UPLC parser's `Constr <tag>` literal now parses via the same
  signed-integer parser as `I`, so the textual syntax can express the
  full domain too.
- Added a PV-matrix test proving: in-range tags behave identically at
  every variant (regression); D/E reject a negative tag as a genuine
  `BuiltinFailure`; A/B/C accept both a negative and a >2^64 tag,
  round-tripping through `unConstrData` as a wide Integer. Added a
  direct CBOR-codec test pinning the exact tag-102 encode bytes for a
  negative and oversized tag and confirming decode-side rejection
  (never a panic).

Mechanical fallout across ~290 call sites in script_context.rs,
populate_v3.rs, populate_v1_v2.rs, populate_gov.rs, tx_info_populate.rs,
eval_redeemer.rs, and the integration test suite (BigInt::from(N)
wrapping at construction sites; ref-binding + guard-based tag checks at
pattern-match sites, since a bare integer literal can no longer be
matched against a BigInt field).

Gates: cargo nextest run -p dugite-uplc (600 passed), cargo clippy -p
dugite-uplc --all-targets -- -D warnings (clean), cargo fmt -p
dugite-uplc -- --check (clean). Also spot-checked cargo check -p
dugite-ledger (the sole consumer of dugite_uplc::phase_two /
tx_info_populate::plutus_data_to_data) still compiles unchanged, since
its own PlutusData::Constr type (dugite-primitives, u64-tagged, wire
decode boundary) is a separate, unaffected type.

+++ crates/dugite-uplc/tests/aiken_v3_always_true.rs
@@ -21,6 +21,7 @@
 //! `#[ignore]` attribute should be removed.
 
 use dugite_uplc::{Constant, Data, Program, Term};
+use num_bigint::BigInt;
 use std::rc::Rc;
 
 /// Inner flat bytes (post-CBOR-unwrap) of the canonical Aiken-built
@@ -66,7 +67,10 @@ fn decode_succeeds() {
 fn cek_reaches_body_no_var_lookup_failure() {
     let p = Program::from_cbor(AIKEN_V3_ALWAYS_TRUE_CBOR)
         .expect("decode canonical Aiken V3 always-true");
-    let stub_ctx = Data::Constr(0, vec![Data::Constr(0, vec![]); 4]);
+    let stub_ctx = Data::Constr(
+        BigInt::from(0),
+        vec![Data::Constr(BigInt::from(0), vec![]); 4],
+    );
     let applied = Term::App(
         Rc::new(p.term),
         Rc::new(Term::Const(Constant::Data(stub_ctx))),
